feat: from_config/from_env directly on Session/SyncSession
Deprecation: standalone from_config/from_env functions.
